msm: mhi_dev: Allocate ring element during M0
Device allocates ring memory during mmio initialization, this
causes 128 event ring elements to be created as MHICFG
register may not be updated by the host. This causes memory
to get exhausted leading to an MMIO init failure.
The Change is to allocate ring memory during M0, during which
MHICFG register is bound to be updated by the host.

This change was earlier reverted to tackle the E911 issue.

Change-Id: Ic3d254affdc663d9c55b04f9230222d8d70412d6
Signed-off-by: Abhinab Abhinab <aabhinab@codeaurora.org>